Renamed the backend endpoint to `experiment`.
