+++ backend/base.py
@@ -1,8 +1,8 @@
 from fastapi import FastAPI
-from backend.routes import submit_router
+from backend.routes import experiment_router
 
 app = FastAPI()
-app.include_router(submit_router)
+app.include_router(experiment_router)
 
 @app.get("/health")
 def health():
